Move failing test to bug-report.test.ts
